requirements.txt
Flask==3.1.1
gunicorn==22.0.0

# Google Sheets
gspread==5.12.0
google-auth==2.29.0
oauth2client==4.1.3

# Image & file handling
pillow==11.2.1

# General utilities
requests==2.32.3
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.1.8
itsdangerous==2.2.0

# Optional for pandas if you're storing extra data
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
